Updated for the tidier returns from RBERT. (#4)
* Updated for the tidier returns from RBERT.

* Bumped version.

* Sync vnum with RBERT for easier cross-matching. For now.

+++ tests/testthat/teardown.R
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-unlink(BERT_PRETRAINED_DIR, recursive = TRUE)
+if (clean_up_cp) unlink(BERT_PRETRAINED_DIR, recursive = TRUE)
